pyproject.toml
[project]
name = "field-image"
version = "0.1.0"
description = "Tools for field imagery organization and post-processing"
readme = "README.md"
requires-python = "==3.12.*"
authors = [
  { name="Eli", email="knodel@umich.edu" }
]
dependencies = [
  "torch==2.7.1",
  "torchvision==0.22.1",
  "pillow==10.4.0",
  "wand==0.6.13",
  "pyyaml==6.0.2",
  "matplotlib==3.10.1",
  "scikit-learn==1.6.1",
  "exifread==3.0.0",
  "pandas==2.2.3",
  "numpy==1.26.4",
  "openpyxl==3.1.5",
  "folium==0.19.7",
  "branca==0.8.1"
]

[project.scripts]
ipro-locator = "FieldImage.apps.iProLocator:main"
create-directory-sites = "FieldImage.apps.createDirectorySites:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]